(maint) Refactoring namespaces
